# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/feedback": {
      "post": {
        "operationId": "FeedbackController_createFeedback",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFeedbackDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Feedback created successfully"
          }
        },
        "summary": "Submit feedback for a doctor",
        "tags": [
          "Feedback"
        ]
      }
    },
    "/feedback/user/{userId}": {
      "get": {
        "operationId": "FeedbackController_getFeedbackByUserId",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of user feedback"
          }
        },
        "summary": "Get all feedback given by a user",
        "tags": [
          "Feedback"
        ]
      }
    },
    "/feedback/{doctorId}": {
      "get": {
        "operationId": "FeedbackController_getFeedbackByDoctorId",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "doctorId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of doctor feedback"
          }
        },
        "summary": "Get all feedback for a specific doctor",
        "tags": [
          "Feedback"
        ]
      }
    },
    "/frontdesk": {
      "get": {
        "operationId": "FrontdeskController_listFrontdesk",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of frontdesk profiles"
          }
        },
        "summary": "List all frontdesk staff",
        "tags": [
          "Frontdesk"
        ]
      },
      "post": {
        "operationId": "FrontdeskController_addFrontdesk",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFrontdeskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Frontdesk profile created"
          },
          "400": {
            "description": "Invalid input or duplicate email"
          }
        },
        "summary": "Add frontdesk staff and create login user",
        "tags": [
          "Frontdesk"
        ]
      }
    },
    "/frontdesk/{userId}": {
      "get": {
        "operationId": "FrontdeskController_getFrontdesk",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "Frontdesk user ID (e.g. FD001)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Frontdesk profile object"
          },
          "404": {
            "description": "Frontdesk profile not found"
          }
        },
        "summary": "Get frontdesk profile by userId",
        "tags": [
          "Frontdesk"
        ]
      },
      "put": {
        "operationId": "FrontdeskController_updateFrontdesk",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "Frontdesk user ID (e.g. FD001)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFrontdeskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Frontdesk profile updated"
          },
          "404": {
            "description": "Frontdesk profile not found"
          }
        },
        "summary": "Edit frontdesk staff by userId",
        "tags": [
          "Frontdesk"
        ]
      }
    },
...
  },
  "info": {
    "title": "MEDBITS Healthcare API",
    "description": "REST API for the MEDBITS Healthcare Management System.\n\n**Authentication**: Pass the user role in the `role` request header.\n\nValid roles: `patient` | `doctor` | `frontdesk` | `admin`",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "role": {
        "type": "apiKey",
        "in": "header",
        "name": "role",
        "description": "User role for RBAC"
      }
    },
    "schemas": {
...
      "CreateFeedbackDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "PAT001"
          },
          "doctorId": {
            "type": "string",
            "example": "DOC001"
          },
          "rating": {
            "type": "string",
            "example": "5"
          },
          "comment": {
            "type": "string",
            "example": "Great doctor!"
          }
        },
        "required": [
          "userId",
          "doctorId",
          "rating"
        ]
      },
      "CreateFrontdeskDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Alice Smith"
          },
          "email": {
            "type": "string",
            "example": "alice@medbits.com"
          },
          "password": {
            "type": "string",
            "example": "password123"
          },
          "phone": {
            "type": "string",
            "example": "1234567890"
          },
          "gender": {
            "type": "string",
            "example": "Female"
          },
          "reportingManagerId": {
            "type": "string",
            "example": "ADM001"
          },
          "languages": {
            "example": [
              "English",
              "Spanish"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "counter": {
            "type": "string",
            "example": "Counter 1"
          },
          "shiftStart": {
            "type": "string",
            "example": "08:00 AM"
          },
          "shiftEnd": {
            "type": "string",
            "example": "04:00 PM"
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ]
      },
      "UpdateFrontdeskDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Alice Smith"
          },
          "email": {
            "type": "string",
            "example": "alice@medbits.com"
          },
          "phone": {
            "type": "string",
            "example": "1234567890"
          },
          "gender": {
            "type": "string",
            "example": "Female"
          },
          "reportingManagerId": {
            "type": "string",
            "example": "ADM001"
          },
          "languages": {
            "example": [
              "English",
              "Spanish"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "counter": {
            "type": "string",
            "example": "Counter 1"
          },
          "shiftStart": {
            "type": "string",
            "example": "08:00 AM"
          },
          "shiftEnd": {
            "type": "string",
            "example": "04:00 PM"
          }
        }
      },
...
    }
  }
}